Refactor to fix Buck Android build
